DarkModeImplementation
ContextAPI created to provider toggle abilities throughout the app. DarkMode changes the logo as well as CSS classes hence the requirement for a context. DarkModeProvider class manipulates the DOM to add CSS classes at the root level.

DarkMode toggle then uses this custom hook that provides the context to change the state of the DarkMode when the button in the header is pressed.