Duration Chart and Sales Chart Implemented
Implementation of duration chart using ReCharts. This component calculates the value each category of stay length has, e.g. how many people stay 1, 2, or 3-4 nights for example.

Implementation of Sales chart using ReCharts. This component creates an object based on the number of days people wish to see on the chart. This object has a label, anad the total sales/extra sales for that day based on the array of bookings passed in. This is then passed into a AreaChart.